pyproject.toml
[project]
name = "commonthread"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "psycopg2-binary>=2.9.10",
    "mkdocs-material>=9.6.12",
    "pyjwt>=2.10.1",
    "pytest>=8.3.5",
    "pytest-django>=4.11.1",
    "python-dotenv>=1.1.0",
    "pygments>=2.19.1",
    "django-cors-headers>=4.7.0",
    "transformers>=4.51.3",
    "openai>=1.78.1",
    "torch>=2.7.0",
    "keybert>=0.9.0",
    "boto3>=1.38.16",
    "django-storages>=1.14.6",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "commonthread.settings"
